+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ###############################
 3.10.1_WIP
 - Fixed -changenick
+- Disabled bot in channel #whatever by default
 
 3.10
 - Added -reverse to reverse your sentences. .secnetnes ruoy esrever ot esrever- deddA