+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ###############################
 #          Changelog          #
 ###############################
-3.10.1_WIP
-- Fixed -changenick
+3.11_WIP
+- Added -update
 - Disabled bot in channel #whatever by default
 - Bot sends all PMs to all valid users (users who are able to control the bot)
+- Fixed -changenick
 
 3.10
 - Added -reverse to reverse your sentences. .secnetnes ruoy esrever ot esrever- deddA